[NFC][GPU] Move reduction configuration to gpu utilities
